feat: assign shared keys to styles

+++ horde/apis/v2/stable.py
@@ -116,6 +116,7 @@ class ImageAsyncGenerate(GenerateTemplate):
 
     def validate(self):
         self.prompt = self.args.prompt
+        self.apikey = self.args.apikey
         self.apply_style()
         super().validate()
         param_validator = ParamValidator(prompt=self.prompt, models=self.args.models, params=self.params, user=self.user)
@@ -244,7 +245,7 @@ class ImageAsyncGenerate(GenerateTemplate):
             r2=self.args.r2,
             shared=shared,
             client_agent=self.args["Client-Agent"],
-            sharedkey_id=self.args.apikey if self.sharedkey else None,
+            sharedkey_id=self.sharedkey.id if self.sharedkey else None,
             proxied_account=self.args["proxied_account"],
             disable_batching=self.args["disable_batching"],
             webhook=self.args.webhook,
@@ -301,8 +302,11 @@ class ImageAsyncGenerate(GenerateTemplate):
                 image_steps=requested_steps,
             )
             if not is_in_limit:
-                self.wp.delete()
-                raise e.BadRequest(fail_message)
+                # If we are using the shared key assigned to a style, then we bypass the shared key requirements
+                # since its owner explicitly allowed to be used with a style exceeding them
+                if not (self.existing_style and self.existing_style.sharedkey and self.existing_style.sharedkey.id == self.sharedkey.id):
+                    self.wp.delete()
+                    raise e.BadRequest(fail_message)
 
     def extrapolate_dry_run_kudos(self):
         self.wp.source_image = self.args.source_image
@@ -363,11 +367,8 @@ class ImageAsyncGenerate(GenerateTemplate):
     def apply_style(self):
         if self.args.style is None:
             return
-        self.existing_style = database.get_style_by_uuid(self.args.style)
-        if not self.existing_style:
-            self.existing_style = database.get_style_by_name(self.args.style)
-        if not self.existing_style:
-            raise e.ThingNotFound("Style", self.args.style)
+        # The super() ensures the common parts of applying a style
+        super().apply_style()
         if self.existing_style.style_type != "image":
             raise e.BadRequest("Text styles cannot be used on image requests", "StyleMismatch")
         if isinstance(self.existing_style, StyleCollection):
@@ -1374,6 +1375,7 @@ class ImageStyle(StyleTemplate):
         code=200,
         description="Lists image styles information",
         as_list=True,
+        skip_none=True,
     )
     def get(self):
         """Retrieves information about all image styles
@@ -1441,6 +1443,7 @@ class ImageStyle(StyleTemplate):
             nsfw=self.args.nsfw,
             prompt=self.args.prompt,
             params=self.args.params if self.args.params is not None else {},
+            sharedkey_id=self.sharedkey.id if self.sharedkey else None,
         )
         new_style.create()
         new_style.set_models(self.models)
@@ -1452,6 +1455,7 @@ class ImageStyle(StyleTemplate):
         }, 200
 
     def validate(self):
+        super().validate()
         if database.get_style_by_name(f"{self.user.get_unique_alias()}::style::{self.style_name}"):
             raise e.BadRequest(
                 (
@@ -1475,6 +1479,7 @@ class SingleImageStyle(SingleStyleTemplate):
         code=200,
         description="Lists image styles information",
         as_list=False,
+        skip_none=True,
     )
     def get(self, style_id):
         """Displays information about an image style."""
@@ -1502,6 +1507,7 @@ class SingleImageStyle(SingleStyleTemplate):
         return super().patch(style_id)
 
     def validate(self):
+        super().validate()
         if (
             self.style_name is not None
             and database.get_style_by_name(f"{self.user.get_unique_alias()}::style::{self.style_name}")
@@ -1545,6 +1551,7 @@ class SingleImageStyleByName(SingleStyleTemplateGet):
         code=200,
         description="Lists image style information by name",
         as_list=False,
+        skip_none=True,
     )
     def get(self, style_name):
         """Seeks an image style by name and displays its information."""
